# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.AttestDeviceRequest:
    properties:
      attestation_data:
        additionalProperties: true
        type: object
      device_id:
        type: string
      device_name:
        type: string
      platform:
        type: string
      spiffe_id:
        type: string
      workload_selector:
        type: string
    required:
    - device_id
    - device_name
    - platform
    type: object
...
  handlers.DeviceResponse:
    properties:
      attestation_data:
        additionalProperties: true
        type: object
      created_at:
        type: string
      device_id:
        type: string
      device_name:
        type: string
      id:
        type: string
      is_verified:
        type: boolean
      platform:
        type: string
      spiffe_id:
        type: string
      trust_level:
        type: integer
      updated_at:
        type: string
      verified_at:
        type: string
      workload_selector:
        type: string
    type: object
...
  handlers.VerifyDeviceRequest:
    properties:
      trust_level:
        maximum: 100
        minimum: 0
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Zero Trust Authentication MVP API documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Zero Trust Auth API
  version: "1.0"
paths:
...
  /devices:
    get:
      consumes:
      - application/json
      description: Get all device attestations for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of devices
          schema:
            items:
              $ref: '#/definitions/handlers.DeviceResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user devices
      tags:
      - devices
    post:
      consumes:
      - application/json
      description: Create a new device attestation for zero trust verification
      parameters:
      - description: Device attestation data
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/handlers.AttestDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created device attestation
          schema:
            $ref: '#/definitions/handlers.DeviceResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Device already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Attest a device
      tags:
      - devices
  /devices/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a device attestation
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Device deleted successfully
        "400":
          description: Invalid device ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Device not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete device
      tags:
      - devices
    get:
      consumes:
      - application/json
      description: Get a specific device attestation by ID
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Device details
          schema:
            $ref: '#/definitions/handlers.DeviceResponse'
        "400":
          description: Invalid device ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Device not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get device by ID
      tags:
      - devices
    put:
      consumes:
      - application/json
      description: Update a device attestation
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated device data
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/handlers.AttestDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated device
          schema:
            $ref: '#/definitions/handlers.DeviceResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Device not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update device
      tags:
      - devices
  /devices/{id}/verify:
    post:
      consumes:
      - application/json
      description: Verify a device attestation and set trust level
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Verification data
        in: body
        name: verification
        required: true
        schema:
          $ref: '#/definitions/handlers.VerifyDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Verified device
          schema:
            $ref: '#/definitions/handlers.DeviceResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - insufficient permissions
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Device not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Verify a device
      tags:
      - devices
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"